requirement.txt
pymongo

python-jsonrpc

pika

requests

redis

sklearn

scipy

lxml

python-dateutil

newspaper

watchdog

ConfigParser

pyyaml